+++ src/JWT/Action/VerifyIdToken/WithLcobucciJWT.php
@@ -17,11 +17,11 @@ use Lcobucci\Clock\FrozenClock;
 use Lcobucci\JWT\Configuration;
 use Lcobucci\JWT\Signer\Key\InMemory;
 use Lcobucci\JWT\Signer\Rsa\Sha256;
-use Lcobucci\JWT\Token as JWT;
+use Lcobucci\JWT\UnencryptedToken;
 use Lcobucci\JWT\Validation\Constraint\IssuedBy;
 use Lcobucci\JWT\Validation\Constraint\PermittedFor;
 use Lcobucci\JWT\Validation\Constraint\SignedWith;
-use Lcobucci\JWT\Validation\Constraint\ValidAt;
+use Lcobucci\JWT\Validation\Constraint\StrictValidAt;
 use Lcobucci\JWT\Validation\ConstraintViolation;
 use Lcobucci\JWT\Validation\RequiredConstraintsViolated;
 use Throwable;
@@ -51,12 +51,9 @@ final class WithLcobucciJWT implements Handler
 
         try {
             $token = $this->config->parser()->parse($tokenString);
+            \assert($token instanceof UnencryptedToken);
         } catch (Throwable $e) {
             throw IdTokenVerificationFailed::withTokenAndReasons($tokenString, ['The token is invalid', $e->getMessage()]);
-        }
-
-        if (!($token instanceof JWT\Plain)) {
-            throw IdTokenVerificationFailed::withTokenAndReasons($tokenString, ['The token could not be decrypted']);
         }
 
         $key = $this->getKey($token);
@@ -66,7 +63,7 @@ final class WithLcobucciJWT implements Handler
 
         try {
             $this->config->validator()->assert($token, ...[
-                new ValidAt($clock, $leeway),
+                new StrictValidAt($clock, $leeway),
                 new IssuedBy(...["https://securetoken.google.com/{$this->projectId}"]),
                 new PermittedFor($this->projectId),
                 new SignedWith($this->config->signer(), InMemory::plainText($key)),
@@ -107,7 +104,7 @@ final class WithLcobucciJWT implements Handler
         return TokenInstance::withValues($tokenString, $headers, $claims);
     }
 
-    private function getKey(JWT $token): string
+    private function getKey(UnencryptedToken $token): string
     {
         if (empty($keys = $this->keys->all())) {
             throw IdTokenVerificationFailed::withTokenAndReasons($token->toString(), ["No keys are available to verify the token's signature."]);
@@ -122,7 +119,7 @@ final class WithLcobucciJWT implements Handler
         throw IdTokenVerificationFailed::withTokenAndReasons($token->toString(), ["No public key matching the key ID '{$keyId}' was found to verify the signature of this token."]);
     }
 
-    private function assertUserAuthedAt(JWT\Plain $token, DateTimeInterface $now): void
+    private function assertUserAuthedAt(UnencryptedToken $token, DateTimeInterface $now): void
     {
         /** @var int|DateTimeImmutable $authTime */
         $authTime = $token->claims()->get('auth_time');
@@ -144,7 +141,7 @@ final class WithLcobucciJWT implements Handler
         }
     }
 
-    private function assertTenantId(JWT\Plain $token, string $tenantId): void
+    private function assertTenantId(UnencryptedToken $token, string $tenantId): void
     {
         $claim = (array) $token->claims()->get('firebase', []);